Copy changes
Jumping in with an attempt to strengthen the copy a little.

The current implementation felt too focused on "us". Changes are to let the reader be the star of the hero's journey.

Reworded the diversity and openness comment to strengthen the claim I think we are trying to make. It felt a little hollow to say we "believe" in diversity. Like, it exists?